# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SLSEA Solar Generation API",
    "version": "0.1.0",
    "description": "SLSEA Solar Generation API, Stages 1\u20136. Public health and OpenAPI; JWT-protected hierarchy, immutable ingestion, scoped history and operational reads. Current paths are relative to the host root. GET/HEAD business representations support authorized ETag revalidation. Deployment and district summaries remain planned."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current API host"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/installations/{installationId}/readings": {
      "post": {
        "operationId": "appendReading",
        "tags": [
          "Readings"
        ],
        "summary": "Append a reading for the authenticated installation",
        "description": "Requires readings:write, an active installation principal matching the URL, and current credential version. Unknown fields and excessive precision return 40002. Repeated timestamps return 40901; inconsistent cumulative counters return 40902. Per-installation transaction locks serialize API ingestion. Meter resets are unsupported. Devices receive the created representation but cannot GET Location with a device token. GET history is available to analysts; Allow is GET, HEAD, POST, OPTIONS. Conditional headers are supported on GET/HEAD, not this ingestion operation.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "installationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created immutable observation. Location and Content-Location identify the analyst-readable canonical resource.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Canonical /readings/{readingId} URI."
              },
              "Content-Location": {
                "schema": {
                  "type": "string"
                },
                "description": "URI of this representation."
              },
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Last-Modified": {
                "$ref": "#/components/headers/RepresentationModified"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reading"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "409": {
            "$ref": "#/components/responses/ReadingConflict"
          }
        }
      },
      "get": {
        "operationId": "getInstallationReadings",
        "tags": [
          "Readings"
        ],
        "summary": "Read authorized historical observations",
        "description": "Requires installation:read and an active analyst. Count and page use one REPEATABLE READ snapshot. Unknown/repeated parameters return 400. Empty results and beyond-end offsets return 200. Regional hidden/nonexistent filters produce an empty page; a hidden/nonexistent nested installation returns 404. Offset pages can shift between requests; a fixed time window still permits late-arriving backfill. GET supports bodyless HEAD. Hierarchy resources use ETags; no Last-Modified is invented for directories lacking a complete membership-change clock.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "installationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "province-id",
            "in": "query",
            "description": "Optional exact UUID filter, ANDed with every other filter and the authenticated jurisdiction.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "district-id",
            "in": "query",
            "description": "Optional exact UUID filter, ANDed with every other filter and the authenticated jurisdiction.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "substation-id",
            "in": "query",
            "description": "Optional exact UUID filter, ANDed with every other filter and the authenticated jurisdiction.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "installation-id",
            "in": "query",
            "description": "Optional exact UUID filter, ANDed with every other filter and the authenticated jurisdiction.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive observation-time bound. Include a timezone; encode + as %2B. At most three fractional digits. from must be before to.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive observation-time bound. Include a timezone; encode + as %2B. At most three fractional digits. from must be before to.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "timestamp",
                "-timestamp"
              ],
              "default": "-timestamp"
            },
            "description": "Timestamp followed by UUID tie-breaker, both in the selected direction."
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 2147483647,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Strong comparison; supports quoted tag lists and *. Evaluated before other preconditions. Weak tags never match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Weak comparison; supports quoted tag lists and *. Matching authorized GET/HEAD returns 304. Takes precedence over If-Modified-Since.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date precondition where a modification timestamp is available. Ignored when If-Match is present. Subsecond comparisons are conservative.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date; ignored when If-None-Match is present. Date-only 304 is used only for immutable atomic readings; mutable views revalidate by ETag to avoid same-second/late-commit false cache hits.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized filtered page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingPage"
                }
              }
            },
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              },
              "Last-Modified": {
                "$ref": "#/components/headers/RepresentationModified"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "304": {
            "description": "Authorized representation unchanged; no body. Validators and cache headers are retained. Empty histories may omit Last-Modified.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              },
              "Last-Modified": {
                "$ref": "#/components/headers/RepresentationModified"
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          }
        }
      }
    },
    "/readings/{readingId}": {
      "get": {
        "operationId": "getReading",
        "tags": [
          "Readings"
        ],
        "summary": "Retrieve an immutable reading",
        "description": "Requires installation:read and an active analyst whose current stored role and jurisdiction match the JWT. Out-of-scope and absent readings return 404. HEAD returns headers without a body; OPTIONS advertises GET, HEAD, OPTIONS. PUT, PATCH and DELETE return 405. GET supports bodyless HEAD. Hierarchy resources use ETags; no Last-Modified is invented for directories lacking a complete membership-change clock.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "readingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Strong comparison; supports quoted tag lists and *. Evaluated before other preconditions. Weak tags never match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Weak comparison; supports quoted tag lists and *. Matching authorized GET/HEAD returns 304. Takes precedence over If-Modified-Since.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date precondition where a modification timestamp is available. Ignored when If-Match is present. Subsecond comparisons are conservative.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date; ignored when If-None-Match is present. Date-only 304 is used only for immutable atomic readings; mutable views revalidate by ETag to avoid same-second/late-commit false cache hits.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized representation.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              },
              "Last-Modified": {
                "$ref": "#/components/headers/RepresentationModified"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reading"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "304": {
            "description": "Authorized representation unchanged; no body. Validators and cache headers are retained. Empty histories may omit Last-Modified.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              },
              "Last-Modified": {
                "$ref": "#/components/headers/RepresentationModified"
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          }
        }
      }
    },
    "/installations/{installationId}/latest-reading": {
      "get": {
        "operationId": "getLatestReading",
        "tags": [
          "Readings"
        ],
        "summary": "Retrieve the latest observation",
        "description": "Latest is selected by observation timestamp and id tie-breaker. Late arrivals cannot replace a newer observation. Empty, absent and hidden installations return 404. Latest does not imply fresh or live. HEAD returns headers without a body; OPTIONS advertises GET, HEAD, OPTIONS. PUT, PATCH and DELETE return 405. GET supports bodyless HEAD. Hierarchy resources use ETags; no Last-Modified is invented for directories lacking a complete membership-change clock.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "installationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Strong comparison; supports quoted tag lists and *. Evaluated before other preconditions. Weak tags never match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Weak comparison; supports quoted tag lists and *. Matching authorized GET/HEAD returns 304. Takes precedence over If-Modified-Since.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date precondition where a modification timestamp is available. Ignored when If-Match is present. Subsecond comparisons are conservative.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date; ignored when If-None-Match is present. Date-only 304 is used only for immutable atomic readings; mutable views revalidate by ETag to avoid same-second/late-commit false cache hits.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized representation.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              },
              "Last-Modified": {
                "$ref": "#/components/headers/RepresentationModified"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reading"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "304": {
            "description": "Authorized representation unchanged; no body. Validators and cache headers are retained. Empty histories may omit Last-Modified.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              },
              "Last-Modified": {
                "$ref": "#/components/headers/RepresentationModified"
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          }
        }
      }
    },
    "/installations/{installationId}/overview": {
      "get": {
        "operationId": "getInstallationOverview",
        "tags": [
          "Readings"
        ],
        "summary": "Retrieve installation, hierarchy and latest reading",
        "description": "One scoped SQL snapshot returns safe installation metadata, ancestor references and lastKnownReading (null for an empty site). No unbounded history or credentials are returned. HEAD returns headers without a body; OPTIONS advertises GET, HEAD, OPTIONS. PUT, PATCH and DELETE return 405. GET supports bodyless HEAD. Hierarchy resources use ETags; no Last-Modified is invented for directories lacking a complete membership-change clock.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "installationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Strong comparison; supports quoted tag lists and *. Evaluated before other preconditions. Weak tags never match.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "Weak comparison; supports quoted tag lists and *. Matching authorized GET/HEAD returns 304. Takes precedence over If-Modified-Since.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Unmodified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date precondition where a modification timestamp is available. Ignored when If-Match is present. Subsecond comparisons are conservative.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "required": false,
            "description": "HTTP-date; ignored when If-None-Match is present. Date-only 304 is used only for immutable atomic readings; mutable views revalidate by ETag to avoid same-second/late-commit false cache hits.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized representation.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              },
              "Last-Modified": {
                "$ref": "#/components/headers/RepresentationModified"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallationOverview"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "304": {
            "description": "Authorized representation unchanged; no body. Validators and cache headers are retained. Empty histories may omit Last-Modified.",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/RepresentationTag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateRevalidation"
              },
              "Vary": {
                "schema": {
                  "type": "string"
                },
                "description": "Includes Accept and Authorization."
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              },
              "Last-Modified": {
                "$ref": "#/components/headers/RepresentationModified"
              }
            }
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          }
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "RequestId": {
        "description": "A server-generated UUID for correlating this response with application logs. For errors, it matches error.requestId.",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "example": "7d50a2a4-39a5-46cc-a8ae-a6ccba1d3897"
      },
      "NoStore": {
        "description": "Foundation health, documentation and error responses must not be stored by caches.",
        "required": true,
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      },
...
      "RepresentationTag": {
        "description": "Strong SHA-256 validator of the serialized representation. Mutable views include URL and jurisdiction context. Authorization is always checked before preconditions.",
        "schema": {
          "type": "string",
          "pattern": "^\"[a-f0-9]{64}\"$"
        }
      },
      "RepresentationModified": {
        "description": "HTTP-date from stored receipt/modification times, not observation time. Milliseconds are retained internally for conservative date comparisons.",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
...
      "ReadingInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "timestamp",
          "powerKw",
          "cumulativeEnergyKwh",
          "voltage"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?(?:Z|[+-]\\d{2}:\\d{2})$",
            "description": "Seconds and timezone required, at most three fractional digits. On or after Unix epoch and commissioning date in Asia/Colombo; at most five minutes ahead of server time. Backfill must fit both counter neighbours."
          },
          "powerKw": {
            "type": "number",
            "minimum": 0,
            "maximum": 999999999.999,
            "multipleOf": 0.001
          },
          "cumulativeEnergyKwh": {
            "type": "number",
            "minimum": 0,
            "maximum": 999999999999.999,
            "multipleOf": 0.001
          },
          "voltage": {
            "type": "number",
            "minimum": 0,
            "maximum": 999999.999,
            "multipleOf": 0.001
          }
        },
        "example": {
          "timestamp": "2026-09-23T06:00:00Z",
          "powerKw": 3.42,
          "cumulativeEnergyKwh": 1267.84,
          "voltage": 231.2
        }
      },
...
    },
    "responses": {
      "InvalidRequest": {
        "description": "40001: The request is invalid, including malformed JSON.",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Missing, invalid or expired credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string",
              "const": "Bearer"
            },
            "description": "Bearer authentication challenge."
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated principal lacks the required scope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived HS256 token issued by POST /auth/token."
      }
    }
  }
}